Move component StaticCard from 'components' to 'components/_shared'
