+++ components/_shared/StaticCard.vue
